requirements.txt
pandas
logging
gspread
scikit-learn
random
google-auth
google-auth-oauthlib
google-api-python-client